+++ examples/asa/scripts/transfer/common.js
@@ -2,9 +2,11 @@ const { types } = require("@algo-builder/web");
 
 exports.tryExecuteTx = async function (deployer, txnParams) {
 	try {
-		await deployer.executeTx(txnParams);
+		const txnParameters = Array.isArray(txnParams) ? txnParams : [txnParams];
+		return await deployer.executeTx(txnParameters);
 	} catch (e) {
 		console.error("Transaction Failed", e.response ? e.response.error : e);
+		throw e;
 	}
 };