+++ examples/permissioned-token/scripts/1-setup-controller.js
@@ -20,22 +20,26 @@ async function setupControllerSSC(runtimeEnv, deployer) {
 	};
 
 	console.log("\n** Deploying smart contract: controller **");
-	const controllerAppInfo = await deployer.deployApp(
-		owner,
-		{
-			appName: "Controller",
-			metaType: types.MetaType.FILE,
-			approvalProgramFilename: "controller.py", // approval program
-			clearProgramFilename: "clear_state_program.py", // clear program
-			localInts: 0,
-			localBytes: 0,
-			globalInts: 2, // 1 to store kill_status, 1 for storing permissions_app_id
-			globalBytes: 0,
-			foreignAssets: [tesla.assetIndex], // pass token_id in foreign assets array
-		},
-		{},
-		templateParam
-	); // pass token_id as a template paramenter
+	const controllerAppInfo = await deployer
+		.deployApp(
+			owner,
+			{
+				appName: "Controller",
+				metaType: types.MetaType.FILE,
+				approvalProgramFilename: "controller.py", // approval program
+				clearProgramFilename: "clear_state_program.py", // clear program
+				localInts: 0,
+				localBytes: 0,
+				globalInts: 2, // 1 to store kill_status, 1 for storing permissions_app_id
+				globalBytes: 0,
+				foreignAssets: [tesla.assetIndex], // pass token_id in foreign assets array
+			},
+			{},
+			templateParam
+		) // pass token_id as a template paramenter
+		.catch((error) => {
+			throw error;
+		});
 	console.log(controllerAppInfo);
 
 	// Use executeSignedTxnFromFile function to execute deployment transaction from a signed file: