+++ packages/runtime/src/errors/errors-list.ts
@@ -210,7 +210,7 @@ by an index that does not exist.`,
 	MAX_LEN_EXCEEDED: {
 		number: 1030,
 		message:
-			"Length of provided TEAL code = %length% exceeds max length of %maxlen%, Mode: %mode%",
+			"Length of provided TEAL code = %length% bytes exceeds max length of %maxlen% bytes, Mode: %mode%",
 		title: "MaxLength Error",
 		description: `MaxLength Error`,
 	},
@@ -456,8 +456,39 @@ maximun uint128`,
 		description:
 			"Maximum program length %maxAppProgramLen% exceeded. Use `ApprovalProgramPages` instead",
 	},
-	UNKNOWN_BLOCK_FIELD: {
+	UNSUPPORTED_VRF_STANDARD: {
 		number: 1067,
+		message: "vtf_standard not supported",
+		title: "Unsupported by runtime",
+		description: "vrf_standard field is not supported yet by algorand. Use vrf_algorand instead",
+	},
+	INVALID_PUB_KEY_LENGTH: {
+		number: 1068,
+		message: "Wrong publicKey size",
+		title: "Wrong publicKey size",
+		description: "Wrong publicKey size. Got: %length%, expected 32",
+	},
+	INVALID_PROOF_LENGTH: {
+		number: 1069,
+		message: "Wrong proof size",
+		title: "Wrong proof size",
+		description: "Wrong proof size. Got: %length%, expected 80",
+	},
+	UNKNOWN_VRF_TYPE: {
+		number: 1070,
+		message: "Unknown vrf_verify field",
+		title: "Unknown field",
+		description: "Unknow field for vrf_verify opcode",
+	},
+	EXTRA_PAGES_EXCEEDED: {
+		number: 1071,
+		message: "Allowed extra pages range is from 0 to %maxExtraAppProgramPages% but got %extraPages% extra pages",
+		title: "Extra pages",
+		description:
+			"Allowed extra pages range is from 0 to %maxExtraAppProgramPages% but got %extraPages% extra pages",
+	},
+	UNKNOWN_BLOCK_FIELD: {
+		number: 1072,
 		message: "Unknown block opcode field",
 		title: "Unknown field",
 		description: "Block opcode support only BlkTimestamp and BlkSeed field, got: %field",